More revisions to authentication. Add JWTAuth.

+++ app/Http/Controllers/MeetingController.php
@@ -8,13 +8,14 @@ use Illuminate\Http\Request;
 use App\Http\Requests;
 
 use App\Meeting;
-// use JWTAuth;
+use JWTAuth;
 
 class MeetingController extends Controller
 {
 	public function __construct(){
-		// middleware to come;
-		// $this->middleware('name');
+		// this would protect everything
+		// $this->middleware('jwt.auth');
+		$this->middleware('jwt.auth', ['only' => ['store','update', 'destroy']]);
 	}
 
 	/**
@@ -53,14 +54,17 @@ class MeetingController extends Controller
 		$this->validate($request, [
 			'title' => 'required',
 			'description' => 'required',
-			'time' => 'required|date_format:YmdHie',
-			'user_id' => 'required'
+			'time' => 'required|date_format:YmdHie'
 		]);
+
+		if (!$user = JWTAuth::parseToken()->authenticate()) {
+			return response()->json(['msg' => 'User not found'], 404);
+		}
 
 		$title = $request->input('title');
 		$description = $request->input('description');
 		$time = $request->input('time');
-		$user_id = $request->input('user_id');	
+		$user_id = $user->id; // extracted from the token	
 
 		$meeting = new Meeting([
 			'time' => Carbon::createFromFormat('YmdHie', $time),
@@ -126,19 +130,21 @@ class MeetingController extends Controller
 			'title' => 'required',
 			'description' => 'required',
 			'time' => 'required|date_format:YmdHie'
-			// ,'user_id' => 'required'
 		]);
+
+		if (!$user = JWTAuth::parseToken()->authenticate()) {
+			return response()->json(['msg' => 'User not found'], 404);
+		}
 
 		$title = $request->input('title');
 		$description = $request->input('description');
 		$time = $request->input('time');
-		// $user_id = $user->id;
-		$user_id = $request->input('user_id');
+		$user_id = $user->id;
 
 		$meeting = Meeting::with('users')->findOrFail($id);
 		
 		if (!$meeting->users()->where('users.id', $user_id)->first()) {
-			return response()->json(['msg' => 'user not registered for the meeting, update not successful'], 401);
+			return response()->json(['msg' => 'User not registered for the meeting, update not successful'], 401);
 		}
 
 		$meeting->time = Carbon::createFromFormat('YmdHie', $time);
@@ -171,6 +177,15 @@ class MeetingController extends Controller
 	public function destroy($id)
 	{
 		$meeting = Meeting::findOrFail($id);
+
+		if (!$user = JWTAuth::parseToken()->authenticate()) {
+			return response()->json(['msg' => 'User not found'], 404);
+		}
+
+		if (!$meeting->users()->where('users.id', $user->id)->first()) {
+			return response()->json(['msg' => 'User not registered for the meeting, update not successful'], 401);
+		}
+
 		$users = $meeting->users();
 		$meeting->users()->detach();
